Renaming the folders.
