+++ ObserverPattern/Observables/IObservable.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ObserverPattern.Observable.Observable
+namespace ObserverPattern.Observables
 {
     interface IObservable
     {
