+++ ObserverPattern/Observer/EmailObserver.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ObserverPattern.Observables;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
